requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
httpx>=0.27
python-multipart>=0.0.9

# ML + data
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4

# File type detection
filetype>=1.2

# OS telemetry
psutil>=5.9
